+++ electron-vite-project/src/vite-env.d.ts
@@ -2,6 +2,7 @@
 
 interface Window {
   electronAPI: {
+    [x: string]: any;
     openExternal(newUrl: string): unknown;
     getIp: () => Promise<{ ip: string; port: string }>;
     getLanguage: () => Promise<string>;